Update scripts for my Chocolatey extensions now set the isFixedVersion to false so the version gets updated on changes
Added cleanup argument in Get-Arguments inside Chocolatey-Package.extension so we can control cleanup of Install files if needed
Added check for cleanup argument in Install-FromZip and Install-Package before deleting files
Fixed Install-WithAutoHotKey to use valida arguments from Get-Arguments and pass those along

+++ Chocolatey-Package.extension/extensions/Installation/Install-FromZip.ps1
@@ -46,6 +46,8 @@ function Install-FromZip {
         Install-ChocolateyInstallPackage @packageArgs
     }
 
-    Get-ChildItem -Path $env:ChocolateyPackageFolder `
-        -Include *.zip, *.7z, *.exe, *.msi, *.reg -Recurse -File | Remove-Item
+    if ($packageArgs.cleanUp) {
+        Get-ChildItem -Path $env:ChocolateyPackageFolder `
+            -Include *.zip, *.7z, *.exe, *.msi, *.reg -Recurse -File | Remove-Item
+    }
 }
